Changed timezone from int to float, changed the lighting a bit.
I made the timezone to a float to be able to fine-tune the sun position (not everyone live at the exact center of a timezone ;-) )
Decreased the specular intensity and shininess, and added a bit of specular light on the continent.

+++ earth.h
@@ -58,7 +58,7 @@ typedef struct _EarthScreen
     
     /* Config parameters */
     float lon, lat;
-    int tz;
+    float tz;
     
     /* Sun position */
     float dec, gha;
